requirements.txt
amqp==5.0.6
asgiref==3.3.1
autopep8==1.5.6
billiard==3.6.4.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-storages==1.11.1
google-api-core==1.26.3
google-auth==1.28.1
google-cloud-core==1.6.0
google-cloud-datastore==1.15.3
google-cloud-ndb==1.8.0
google-cloud-storage==1.37.1
google-cloud-tasks==2.2.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.0
gunicorn==20.1.0
idna==2.10
kombu==5.0.2
libcst==0.3.18
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.9
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.1.2
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
pymemcache==3.4.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webapp2==2.5.2
whitenoise==5.2.0
XlsxWriter==1.3.8
